Added methods `is_integer`, `is_unsigned_integer` and `is_float`, `as_integer`, `as_unsigned_integer` and `as_float` to `json`.

+++ src/jsoncons/jsoncons_config.hpp
@@ -24,7 +24,7 @@
 
 namespace jsoncons {
 
-typedef long double float_type;
+typedef double float_type;
 typedef long long integer_type;
 typedef unsigned long long unsigned_integer_type;
